% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cur_colnames.R
\name{cur_colnames}
\alias{cur_colnames}
\title{Get a vector of current column names}
\usage{
cur_colnames()
}
\description{
A wrapper to \code{tidyselect::peek_vars()}, which is not re-exported in \code{dplyr}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
df <- as.data.frame(as.list(1:10))

identical(
  df,
  df \%>\% select(cur_colnames())
)

df \%>\%
  select(rev(cur_colnames()))

df \%>\%
  select(stringr::str_subset(cur_colnames(), "[1-5]L$"))
}
\seealso{
\code{\link[tidyselect:peek_vars]{tidyselect::peek_vars()}}
}
