% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_cols.R
\name{last_cols}
\alias{last_cols}
\title{Vectorized last_col}
\usage{
last_cols(offsets)
}
\arguments{
\item{offsets}{An integer vector of columns to select from the end.
If a single number is provided, selects \code{(offsets - 1):0}.}
}
\description{
Vectorized last_col
}
\examples{
library(dplyr, warn.conflicts = FALSE)
df <- as.data.frame(as.list(1:10))

# Selects last 5 columns
df \%>\% select(last_cols(5))
# Same as above
df \%>\% select(last_cols(4:0))

# Using tidyselect::last_col()
df \%>\% select(last_col(4):last_col())
}
\seealso{
\code{\link[tidyselect:everything]{tidyselect::last_col()}}
}
