% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_str_sort.R
\name{match_str_sort}
\alias{match_str_sort}
\title{Match and sort columns by name}
\usage{
match_str_sort(match, ...)
}
\arguments{
\item{match}{A regular expression. Passed to \code{tidyselect::matches()}.}

\item{...}{
  Arguments passed on to \code{\link[stringr:str_order]{stringr::str_sort}}
  \describe{
    \item{\code{x}}{A character vector to sort.}
    \item{\code{decreasing}}{A boolean. If \code{FALSE}, the default, sorts from
lowest to highest; if \code{TRUE} sorts from highest to lowest.}
    \item{\code{na_last}}{Where should \code{NA} go? \code{TRUE} at the end,
\code{FALSE} at the beginning, \code{NA} dropped.}
    \item{\code{locale}}{Locale to use for comparisons. See
\code{\link[stringi:stri_locale_list]{stringi::stri_locale_list()}} for all possible options.
Defaults to "en" (English) to ensure that default behaviour is
consistent across platforms.}
    \item{\code{numeric}}{If \code{TRUE}, will sort digits numerically, instead
of as strings.}
  }}
}
\description{
Sorts a regex-matched set of columns using \code{stringr::str_sort()}
}
\examples{
library(dplyr, warn.conflicts = FALSE)
df <- data.frame(c = 1, ch = 3, h = 2, dont_pick = 0)

df \%>\%
  select(match_str_sort("^[a-z]+$"))

# "ch" sorts after "h" in Czech locale
df \%>\%
  select(match_str_sort("^[a-z]+$", locale = "cs"))
}
\seealso{
\code{\link[stringr:str_order]{stringr::str_sort()}}, \code{\link[tidyselect:starts_with]{tidyselect::matches()}}
}
