% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_n_of.R
\name{only_n_of}
\alias{only_n_of}
\title{Select exact number of columns from a character vector}
\usage{
only_n_of(x, n, ..., vars = NULL)
}
\arguments{
\item{x}{A vector of character names or numeric locations.}

\item{n}{Exact number of column matches to expect}

\item{...}{Unused.}

\item{vars}{A character vector of variable names. If not supplied,
the variables are taken from the current selection context (as
established by functions like \code{select()} or \code{pivot_longer()}).}
}
\description{
Select exact number of columns from a character vector
}
\examples{
library(dplyr, warn.conflicts = FALSE)
df <- data.frame(ID = 1, id = 1)
id_vec <- c("ID", "id")

df \%>\% select(any_of(id_vec))

df \%>\% select(only_n_of(id_vec, 2))
try(
  df \%>\% select(only_n_of(id_vec, 1))
)
}
